requirements.txt
kafka-python
pymongo
streamlit